+++ bioconda_utils/build.py
@@ -8,12 +8,17 @@ import os
 import logging
 import itertools
 
-from typing import List
+from typing import List, Optional
+from bioconda_utils.skiplist import Skiplist
+from bioconda_utils.build_failure import BuildFailureRecord
+from bioconda_utils.githandler import GitHandler
 
+import conda
 from conda.exports import UnsatisfiableError
 from conda_build.exceptions import DependencyNeedsBuildingError
 import networkx as nx
 import pandas
+from ruamel_yaml import YAML
 
 from . import utils
 from . import docker_utils
@@ -51,7 +56,10 @@ def build(recipe: str, pkg_paths: List[str] = None,
           docker_builder: docker_utils.RecipeBuilder = None,
           raise_error: bool = False,
           linter=None,
-          mulled_conda_image: str = pkg_test.MULLED_CONDA_IMAGE) -> BuildResult:
+          mulled_conda_image: str = pkg_test.MULLED_CONDA_IMAGE,
+          record_build_failure: bool = False,
+          dag: Optional[nx.DiGraph] = None,
+          skiplist_leafs: bool = False) -> BuildResult:
     """
     Build a single recipe for a single env
 
@@ -68,7 +76,13 @@ def build(recipe: str, pkg_paths: List[str] = None,
       raise_error: Instead of returning a failed build result, raise the
         error instead. Used for testing.
       linter: Linter to use for checking recipes
+      record_build_failure: If True, record build failures in a file next to the meta.yaml
+      dag: optional nx.DiGraph with dependency information
+      skiplist_leafs: If True, blacklist leaf packages that fail to build
     """
+    if record_build_failure and not dag:
+        raise ValueError("record_build_failure requires dag to be set")
+
     if linter:
         logger.info('Linting recipe %s', recipe)
         linter.clear_messages()
@@ -141,6 +155,8 @@ def build(recipe: str, pkg_paths: List[str] = None,
 
     except (docker_utils.DockerCalledProcessError, sp.CalledProcessError) as exc:
         logger.error('BUILD FAILED %s', recipe)
+        if record_build_failure:
+            store_build_failure(recipe, exc.output, meta, dag, skiplist_leafs)
         if raise_error:
             raise exc
         return BuildResult(False, None)
@@ -160,6 +176,24 @@ def build(recipe: str, pkg_paths: List[str] = None,
         return BuildResult(True, mulled_images)
 
     return BuildResult(True, None)
+
+
+def store_build_failure(recipe, output, meta, dag, skiplist_leafs):
+    """
+    Write the exception to a file next to the meta.yaml
+    """
+    pkg_name = meta["package"]["name"]
+    is_leaf = dag.out_degree(pkg_name) == 0
+
+    build_failure_record = BuildFailureRecord(recipe)
+    build_failure_record.set_commit_sha_to_current_recipe()
+    # if recipe is a leaf (i.e. not used by others as dependency)
+    # we can automatically blacklist it if desired
+    build_failure_record.skiplist = skiplist_leafs and is_leaf
+    build_failure_record.log = output.decode("utf-8")
+
+    logger.info(f"Storing build failure record for recipe {recipe}")
+    build_failure_record.git_handler.commit_and_push_changes([build_failure_record.path], None, f"Add build failure record for recipe {recipe}")
 
 
 def remove_cycles(dag, name2recipes, failed, skip_dependent):
@@ -228,7 +262,9 @@ def build_recipes(recipe_folder: str, config_path: str, recipes: List[str],
                   n_workers: int = 1,
                   worker_offset: int = 0,
                   keep_old_work: bool = False,
-                  mulled_conda_image: str = pkg_test.MULLED_CONDA_IMAGE):
+                  mulled_conda_image: str = pkg_test.MULLED_CONDA_IMAGE,
+                  record_build_failures: bool = False,
+                  skiplist_leafs: bool = False):
     """
     Build one or many bioconda packages.
 
@@ -261,7 +297,7 @@ def build_recipes(recipe_folder: str, config_path: str, recipes: List[str],
         return True
 
     config = utils.load_config(config_path)
-    blacklist = utils.get_blacklist(config, recipe_folder)
+    blacklist = Skiplist(config, recipe_folder)
 
     # get channels to check
     if check_channels is None:
@@ -305,7 +341,6 @@ def build_recipes(recipe_folder: str, config_path: str, recipes: List[str],
                for package in nx.topological_sort(subdag)
                for recipe in name2recipes[package]]
 
-
     built_recipes = []
     skipped_recipes = []
     failed_uploads = []
@@ -347,7 +382,10 @@ def build_recipes(recipe_folder: str, config_path: str, recipes: List[str],
                     channels=config['channels'],
                     docker_builder=docker_builder,
                     linter=linter,
-                    mulled_conda_image=mulled_conda_image)
+                    mulled_conda_image=mulled_conda_image,
+                    dag=dag,
+                    record_build_failure=record_build_failures,
+                    skiplist_leafs=skiplist_leafs)
 
         if not res.success:
             failed.append(recipe)